Randomized the creation of datasets
Before, the first x amount of molecules in each page of the ANI-1x h5 dataset were used for training. Now, you can set the percentage of of each page you want to use for the training dataset, and it takes a random percent of indices from the page and uses those.

I also made "1_adjust_energy_scales" also adjust the energy scales for nitrogen.